Fix the Ladders, Stoves, and Counters not being fully painted
* Textures/Tiles/Kitchen/StovesPipe.png
* Textures/Tiles/Kitchen/Stoves.png:
Merge into a single file to simplify the painting process.

* Tiles/Kitchen/Stoves.cs (PostDraw): Get both textures using
`TileDrawing.GetTileDrawTexture', so that painting works.
Use the values from `TileDrawing.GetTileDrawData' to draw the
tile on the stovetop, to avoid inconsistencies.

* Textures/Tiles/Kitchen/CountersAlign.png:
* Textures/Tiles/Kitchen/Counters.png:
Merge into a single file to simplify the painting process.
Put all counters on the top row, so that they may be stood on.

* Tiles/Kitchen/Counters.cs (SetStaticDefaults): Don't wrap styles.
(KillMultiTile): Take into account the style changes.
(PostDraw): Take into account the style changes.
Get texture using `TileDrawing.GetTileDrawTexture', so that painting works.

* Tiles/Ladder.cs (PostDraw): Get texture using
`TileDrawing.GetTileDrawTexture', so that painting works.

* Tiles/Plastic/Food/Eggnog.cs: Remove outdated commented-out code.