Clean up Rubblemaker items and make them functional
* CFUHooks.cs: Set up `GetFlexibleTileWand' hook and
selectively uncomment the rubblemaker tilewands.

*  CFUVanillaOverrides.cs (GlobalCFUItem):
Replace the placeholder rubblemaker as soon as its created.

* Items/Rubblemaker/RubblemakerPileLarge.cs:
* Items/Rubblemaker/RubblemakerPileMedium.cs:
* Items/Rubblemaker/RubblemakerPileSmall.cs:
* Items/Rubblemaker/RubblemakerPileStalactiteMedium.cs:
* Items/Rubblemaker/RubblemakerPileStalactiteSmall.cs:
* Items/Rubblemaker/RubblemakerPlaceholder.cs:
* Items/Rubblemaker/RubblemakerPilePot.cs:
Remove "pile" from the filenames where it didn't apply.
Replace "StalactiteMedium" and "StalactiteSmall"
into just "Stalactite" and "Stalagmite".
Add item rarities and +3 range boost to mimic vanilla.
Allow for switching the type while holding the item
(as opposed to only in the inventory).
Draw the charge bar in the inventory, same as vanilla.
Trim down descriptions so they only refer to the currently
selected type.

* Textures/Items/Rubblemaker/RubblemakerStalactiteMedium.png:
Rename to just Stalactite and change color to cyan.

* Textures/Items/Rubblemaker/RubblemakerStalactiteMedium.png:
Rename to just Stalagmite and change color to purple.

* Textures/Items/Rubblemaker/RubblemakerPot.png:
Change color to lime.

* Textures/Items/Rubblemaker/RubblemakerBar.png: New file.